型が多いのでファイルを分ける。未実装と勘違いする容易になってた
