+++ src/rmmz/eventCommand/commands/flow/branch/types/index.ts
@@ -1,6 +1,9 @@
 export * from "./actor";
 export * from "./branch";
 export * from "./branchParams";
+export * from "./button";
+export * from "./character";
 export * from "./command";
 export * from "./else";
+export * from "./enemy";
 export * from "./variable";
